add functionallity for add, edit and delete task
