+++ src/modules/tasks.js
@@ -80,30 +80,4 @@ class Tasks {
   }
 }
 
-export const editText = () => {
-  const allTexts = document.querySelectorAll('.input_text');
-
-  const changeValue = (e) => {
-    const id = e.target.id.split('_')[0];
-    const text = document.getElementById(`${id}_text`);
-    if (e.key === 'Enter') {
-      tasksObject.taskList.forEach((element) => {
-        if (Number(id) === element.index) {
-          element.description = text.value;
-        }
-      });
-      dataStore();
-      tasks.innerHTML = '';
-      tasksObject.taskList.forEach((element) => {
-        const { index, description } = element;
-        tasksObject.display(index, description);
-        editText();
-        tasksObject.counter += 1;
-      });
-    }
-  };
-
-  allTexts[allTexts.length - 1].addEventListener('keydown', (e) => changeValue(e));
-};
-
 export default Tasks;
